+++ src/Factories/RequestFactoryV25.php
@@ -39,7 +39,7 @@ final class RequestFactoryV25 extends RequestFactoryV2
     {
         $this->authSignatureHandler = new AuthSignatureHandlerV25($keyring);
         $this->userSignatureHandler = new UserSignatureHandlerV2($user, $keyring);
-        $this->orderDataHandler = new OrderDataHandlerV25($bank, $user, $keyring);
+        $this->orderDataHandler = new OrderDataHandlerV25($user, $keyring);
         $this->digestResolver = new DigestResolverV2();
         parent::__construct($bank, $user, $keyring);
     }